Revert "search ui: search results components hierarchy refactor (#42483)"
This reverts commit 7c2e6248da0f6aaf4c84e6c8a51c62f55e9a564f.
