+++ client/search-ui/src/components/RepoSearchResult.tsx
@@ -2,6 +2,7 @@ import React, { useEffect, useRef } from 'react'
 
 import { mdiSourceFork, mdiArchive, mdiLock } from '@mdi/js'
 import classNames from 'classnames'
+import SourceRepositoryIcon from 'mdi-react/SourceRepositoryIcon'
 
 import { highlightNode } from '@sourcegraph/common'
 import { codeHostSubstrLength, displayRepoName } from '@sourcegraph/shared/src/components/RepoLink'
@@ -33,13 +34,85 @@ export const RepoSearchResult: React.FunctionComponent<RepoSearchResultProps> = 
     const repoDescriptionElement = useRef<HTMLDivElement>(null)
     const repoNameElement = useRef<HTMLAnchorElement>(null)
 
-    const title = (
+    const renderTitle = (): JSX.Element => (
         <div className={styles.title}>
             <span className={classNames('test-search-result-label', styles.titleInner, styles.mutedRepoFileLink)}>
                 <Link to={getRepoMatchUrl(result)} ref={repoNameElement}>
                     {displayRepoName(getRepoMatchLabel(result))}
                 </Link>
             </span>
+        </div>
+    )
+
+    const renderBody = (): JSX.Element => (
+        <div data-testid="search-repo-result">
+            <div className={classNames(styles.searchResultMatch, 'p-2 flex-column')}>
+                {result.repoLastFetched && <LastSyncedIcon lastSyncedTime={result.repoLastFetched} />}
+                <div className="d-flex align-items-center flex-row">
+                    <div className={styles.matchType}>
+                        <small>Repository match</small>
+                    </div>
+                    {result.fork && (
+                        <>
+                            <div className={styles.divider} />
+                            <div>
+                                <Icon
+                                    aria-label="Forked repository"
+                                    className={classNames('flex-shrink-0 text-muted', styles.icon)}
+                                    svgPath={mdiSourceFork}
+                                />
+                            </div>
+                            <div>
+                                <small>Fork</small>
+                            </div>
+                        </>
+                    )}
+                    {result.archived && (
+                        <>
+                            <div className={styles.divider} />
+                            <div>
+                                <Icon
+                                    aria-label="Archived repository"
+                                    className={classNames('flex-shrink-0 text-muted', styles.icon)}
+                                    svgPath={mdiArchive}
+                                />
+                            </div>
+                            <div>
+                                <small>Archived</small>
+                            </div>
+                        </>
+                    )}
+                    {result.private && (
+                        <>
+                            <div className={styles.divider} />
+                            <div>
+                                <Icon
+                                    aria-label="Private repository"
+                                    className={classNames('flex-shrink-0 text-muted', styles.icon)}
+                                    svgPath={mdiLock}
+                                />
+                            </div>
+                            <div>
+                                <small>Private</small>
+                            </div>
+                        </>
+                    )}
+                </div>
+                {result.description && (
+                    <>
+                        <div className={styles.dividerVertical} />
+                        <div>
+                            <small>
+                                <em ref={repoDescriptionElement}>
+                                    {result.description.length > REPO_DESCRIPTION_CHAR_LIMIT
+                                        ? result.description.slice(0, REPO_DESCRIPTION_CHAR_LIMIT) + ' ...'
+                                        : result.description}
+                                </em>
+                            </small>
+                        </div>
+                    </>
+                )}
+            </div>
         </div>
     )
 
@@ -75,83 +148,17 @@ export const RepoSearchResult: React.FunctionComponent<RepoSearchResultProps> = 
     return (
         <ResultContainer
             index={index}
-            title={title}
+            icon={SourceRepositoryIcon}
+            collapsible={false}
+            defaultExpanded={true}
+            title={renderTitle()}
             resultType={result.type}
             onResultClicked={onSelect}
+            expandedChildren={renderBody()}
             repoName={result.repository}
             repoStars={result.repoStars}
             className={containerClassName}
             as={as}
-        >
-            <div data-testid="search-repo-result">
-                <div className={classNames(styles.searchResultMatch, 'p-2 flex-column')}>
-                    {result.repoLastFetched && <LastSyncedIcon lastSyncedTime={result.repoLastFetched} />}
-                    <div className="d-flex align-items-center flex-row">
-                        <div className={styles.matchType}>
-                            <small>Repository match</small>
-                        </div>
-                        {result.fork && (
-                            <>
-                                <div className={styles.divider} />
-                                <div>
-                                    <Icon
-                                        aria-label="Forked repository"
-                                        className={classNames('flex-shrink-0 text-muted', styles.icon)}
-                                        svgPath={mdiSourceFork}
-                                    />
-                                </div>
-                                <div>
-                                    <small>Fork</small>
-                                </div>
-                            </>
-                        )}
-                        {result.archived && (
-                            <>
-                                <div className={styles.divider} />
-                                <div>
-                                    <Icon
-                                        aria-label="Archived repository"
-                                        className={classNames('flex-shrink-0 text-muted', styles.icon)}
-                                        svgPath={mdiArchive}
-                                    />
-                                </div>
-                                <div>
-                                    <small>Archived</small>
-                                </div>
-                            </>
-                        )}
-                        {result.private && (
-                            <>
-                                <div className={styles.divider} />
-                                <div>
-                                    <Icon
-                                        aria-label="Private repository"
-                                        className={classNames('flex-shrink-0 text-muted', styles.icon)}
-                                        svgPath={mdiLock}
-                                    />
-                                </div>
-                                <div>
-                                    <small>Private</small>
-                                </div>
-                            </>
-                        )}
-                    </div>
-                    {result.description && (
-                        <>
-                            <div className={styles.dividerVertical} />
-                            <div>
-                                <small>
-                                    <em ref={repoDescriptionElement}>
-                                        {result.description.length > REPO_DESCRIPTION_CHAR_LIMIT
-                                            ? result.description.slice(0, REPO_DESCRIPTION_CHAR_LIMIT) + ' ...'
-                                            : result.description}
-                                    </em>
-                                </small>
-                            </div>
-                        </>
-                    )}
-                </div>
-            </div>
-        </ResultContainer>
+        />
     )
 }
