+++ client/search-ui/src/components/ResultContainer.module.scss
@@ -28,6 +28,11 @@
         flex-wrap: wrap;
     }
 
+    &-divider {
+        border-right: 1px solid var(--border-color-2);
+        height: 1rem;
+    }
+
     p {
         margin-bottom: 0;
     }
@@ -36,12 +41,44 @@
         border-bottom: none;
     }
 
+    &-description {
+        line-height: (14/11);
+    }
+
     :global(.match-highlight) {
         color: var(--text-muted-highlighted);
     }
 }
 
-.result {
+.toggle-matches-container {
+    display: flex;
+    flex-shrink: 0;
+    padding: 0;
+}
+
+.toggle-matches-button {
+    width: 100%;
+    text-align: left;
+    border: none;
+    padding: 0.25rem 0.5rem;
+    background-color: var(--border-color);
+    border-bottom-left-radius: var(--border-radius);
+    border-bottom-right-radius: var(--border-radius);
+    color: var(--text-muted);
+
+    // When expanded, stick collapse button to the bottom of the
+    // screen if there are enough search results to cause scrolling.
+    &--expanded {
+        position: sticky;
+        bottom: 0;
+    }
+
+    &-text {
+        margin-left: 0.125rem;
+    }
+}
+
+.collapsible-results {
     // The LastSyncedIcon is absolutely-positions inside the search results.
     // This causes it to show over the sticky header when scrolling unless
     // we isolate the search result contents.
