+++ client/search-ui/src/components/SearchResult.module.scss
@@ -101,42 +101,7 @@
     }
 }
 
-.header-description {
-    line-height: (14/11);
-}
-
-.toggle-matches-button {
-    width: 100%;
-    text-align: left;
-    border: none;
-    padding: 0.25rem 0.5rem;
-    background-color: var(--border-color);
-    border-bottom-left-radius: var(--border-radius);
-    border-bottom-right-radius: var(--border-radius);
-    color: var(--text-muted);
-
-    // When expanded, stick collapse button to the bottom of the
-    // screen if there are enough search results to cause scrolling.
-    &--expanded {
-        position: sticky;
-        bottom: 0;
-    }
-
-    &-text {
-        margin-left: 0.125rem;
-    }
-}
-
-.collapsible-results {
-    // The LastSyncedIcon is absolutely-positions inside the search results.
-    // This causes it to show over the sticky header when scrolling unless
-    // we isolate the search result contents.
-    isolation: isolate;
-
-    border-radius: var(--border-radius);
-    border: 1px solid var(--border-color);
-
-    &:hover {
-        border-color: var(--border-color-2);
-    }
+.symbols {
+    border: none !important;
+    padding: 0 !important;
 }
