+++ client/search-ui/src/components/index.ts
@@ -2,17 +2,13 @@ export * from './CodeExcerpt'
 export * from './CodeHostIcon'
 export * from './CommitSearchResult'
 export * from './CommitSearchResultMatch'
-export * from './FileContentSearchResult'
+export * from './FileSearchResult'
 export * from './LastSyncedIcon'
 export * from './RepoFileLink'
 export * from './RepoSearchResult'
-export * from './FilePathSearchResult'
-export * from './SymbolSearchResult'
-export * from './LegacyResultContainer'
 export * from './ResultContainer'
 export * from './SearchResultStar'
 export * from './SyntaxHighlightedSearchQuery'
-export * from './codeLinkNavigation'
 
 import FileMatchChildrenStyles from './FileMatchChildren.module.scss'
 import SearchResultStyles from './SearchResult.module.scss'
